Fix/masternode list pipelines

+++ DashSync/shared/Models/Entities/DSBlockchainIdentityKeyPathEntity+CoreDataProperties.h
@@ -21,8 +21,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) uint16_t keyType;
 @property (nonatomic, assign) uint16_t keyStatus;
 @property (nonatomic, assign) uint32_t keyID;
-@property (nonatomic, assign) uint8_t securityLevel;
-@property (nonatomic, assign) uint8_t purpose;
+@property (nonatomic, assign) uint16_t securityLevel;
+@property (nonatomic, assign) uint16_t purpose;
 @property (nullable, nonatomic, retain) NSData *publicKeyData;
 
 @end
